separate has_many clases
