+++ lib/avo/items_holder.rb
@@ -1,0 +1,74 @@
+module Avo
+  class ItemsHolder
+    attr_reader :tools
+    # attr_reader :filters
+    # attr_reader :actions
+    attr_accessor :items
+    attr_accessor :invalid_fields
+
+    def initialize
+      @items = []
+      @items_index = 0
+      @invalid_fields = []
+
+      puts ["ItemsHolder initialize->", self.class].inspect
+    end
+
+    def field(field_name, **args, &block)
+      # puts ["ItemsHolder field->", field_name, resource_class, args, block].inspect
+
+      field_parser = Avo::Dsl::FieldParser.new(id: field_name, order_index: @items_index, **args, &block).parse
+
+      if field_parser.invalid?
+        as = args.fetch(:as, nil)
+
+        # End execution ehre and add the field to the invalid_fileds payload so we know to wanr the developer about that.
+        # @todo: Make sure this warning is still active
+        return add_invalid_field({
+          name: field_name,
+          as: as,
+          # resource: resource_class.name,
+          message: "There's an invalid field configuration for this resource. <br/> <code class='px-1 py-px rounded bg-red-600'>field :#{field_name}, as: #{as}</code>"
+        })
+      end
+
+      add_item field_parser.instance
+    end
+
+    def tabs(instance)
+      add_item instance
+    end
+
+    def tab(name, **args, &block)
+      add_item Avo::TabBuilder.parse_block(name: name, **args, &block)
+    end
+
+    def tool(klass, **args)
+      # puts ["ItemsHolder tool->", klass, args].inspect
+      instance = klass.new **args
+      add_item instance
+    end
+
+    def panel(panel_name = nil, **args, &block)
+      panel = Avo::PanelBuilder.parse_block(name: panel_name, **args, &block)
+
+      add_item panel
+    end
+
+    def add_item(instance)
+      @items << instance
+
+      increment_order_index
+    end
+
+    private
+
+    def add_invalid_field(payload)
+      @invalid_fields << payload
+    end
+
+    def increment_order_index
+      @items_index += 1
+    end
+  end
+end
