+++ lib/avo/main_panel.rb
@@ -1,0 +1,3 @@
+class Avo::MainPanel < Avo::Panel
+  class_attribute :item_type, default: :main_panel
+end
