+++ lib/avo/panel_builder.rb
@@ -1,0 +1,25 @@
+class Avo::PanelBuilder
+  class << self
+    def parse_block(**args, &block)
+      Docile.dsl_eval(new(**args), &block).build
+    end
+  end
+
+  delegate :field, to: :items_holder
+  delegate :items, to: :items_holder
+  # delegate :tool, to: :items_holder
+  # delegate :panel, to: :items_holder
+
+  attr_reader :items_holder
+
+  def initialize(name: nil, **args)
+    @panel = Avo::Panel.new(name: name, **args)
+    @items_holder = Avo::ItemsHolder.new
+  end
+
+  # Fetch the tab
+  def build
+    @panel.items_holder = @items_holder
+    @panel
+  end
+end
